+++ .github/workflows/test_methods.yml
@@ -34,7 +34,7 @@ jobs:
 
     - name: Test FBA config
       run: |
-        python flowsa/test_methods.py
+        python tests/test_methods.py
 
   FBS_testing:
     runs-on: macos-latest
@@ -61,9 +61,9 @@ jobs:
       id: FBS
       if: always() # Proceed even if Test FBA fails
       run: |
-        for m in $(python flowsa/test_FBS_against_remote.py list)
+        for m in $(python tests/test_FBS_against_remote.py list)
         do
-          python flowsa/test_FBS_against_remote.py "$m"
+          python tests/test_FBS_against_remote.py "$m"
         done
 
     - name: Upload csv files
