+++ tests/test_examples.py
